apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.11"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "google-cloud-secret-manager>=2.21.0",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
    "datamodel-code-generator==0.33.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.23.8",
]

[tool.ruff]
exclude = [
    "src/models/**",
]
